refactor: rebase relay package
